app/requirements.txt
Flask==3.0.0
PyPDF2==3.0.0
PyMuPDF==1.23.0
pytesseract==0.3.10
Pillow==10.0.0
opencv-python==4.8.0.76
camelot-py[base]==0.11.0
pandas==2.1.0
python-magic-bin==0.4.14
numpy==1.24.0
werkzeug==3.0.0
